requirements.txt
Faker
mysql.connector